Teste banco apin

+++ conecta-a-vida-api/src/main/java/br/edu/ifsp/conectaavida/repository/UsuarioRepository.java
@@ -5,10 +5,18 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
+/**
+ * INTERFACE: UsuarioRepository
+ * Objetivo: Operar buscas e cadastros na tabela de usuários e administradores ("usuarios").
+ */
 @Repository
-// ATENÇÃO: Segundo parâmetro é Integer porque o ID de Usuário é Integer.
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
-    // Busca customizada para validação de login ou checagem.
+    /**
+     * Procura um usuário através de uma correspondência exata de e-mail.
+     * * Explicação para a Equipa: Como a coluna 'email' possui a restrição UNIQUE no banco do Supabase,
+     * esta consulta é extremamente rápida e assertiva. É o método chave utilizado pelo processo de
+     * login para encontrar o usuário antes de validar se a senha informada está correta.
+     */
     Optional<Usuario> findByEmail(String email);
 }
